requirements.txt
aiohttp==3.5.4
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
click==6.7
colorama==0.4.1
cryptography==2.8
discord.py==1.0.0
docopt==0.6.2
feedparser==5.2.1
gitsome==0.8.4
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gspread==3.1.0
gspread-asyncio==1.0.0
httplib2==0.14.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.5.2
oauth2client==4.1.3
oauthlib==3.1.0
ply==3.11
prompt-toolkit==2.0.10
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.4.2
pylint==2.4.3
pyOpenSSL==19.0.0
python-dotenv==0.10.3
pytz==2016.10
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
uritemplate==3.0.0
uritemplate.py==3.0.2
urllib3==1.25.6
wcwidth==0.1.7
websockets==6.0
wrapt==1.11.2
yarl==1.3.0